Refactor parser compilation units into specific components
