+++ PCbuild/_freeze_module.vcxproj
@@ -163,6 +163,8 @@
     <ClCompile Include="..\Parser\parser.c" />
     <ClCompile Include="..\Parser\peg_api.c" />
     <ClCompile Include="..\Parser\pegen.c" />
+    <ClCompile Include="..\Parser\pegen_errors.c" />
+    <ClCompile Include="..\Parser\action_helpers.c" />
     <ClCompile Include="..\Parser\string_parser.c" />
     <ClCompile Include="..\Parser\token.c" />
     <ClCompile Include="..\Parser\tokenizer.c" />
