+++ PCbuild/pythoncore.vcxproj
@@ -433,6 +433,8 @@
     <ClCompile Include="..\Parser\tokenizer.c" />
     <ClCompile Include="..\Parser\token.c" />
     <ClCompile Include="..\Parser\pegen.c" />
+    <ClCompile Include="..\Parser\pegen_errors.c" />
+    <ClCompile Include="..\Parser\action_helpers.c" />
     <ClCompile Include="..\Parser\parser.c" />
     <ClCompile Include="..\Parser\string_parser.c" />
     <ClCompile Include="..\Parser\peg_api.c" />
