+++ Tools/peg_generator/pegen/build.py
@@ -69,6 +69,8 @@ def compile_c_extension(
                 str(MOD_DIR.parent.parent.parent / "Python" / "asdl.c"),
                 str(MOD_DIR.parent.parent.parent / "Parser" / "tokenizer.c"),
                 str(MOD_DIR.parent.parent.parent / "Parser" / "pegen.c"),
+                str(MOD_DIR.parent.parent.parent / "Parser" / "pegen_errors.c"),
+                str(MOD_DIR.parent.parent.parent / "Parser" / "action_helpers.c"),
                 str(MOD_DIR.parent.parent.parent / "Parser" / "string_parser.c"),
                 str(MOD_DIR.parent / "peg_extension" / "peg_extension.c"),
                 generated_source_path,
